# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API сервиса подачи заявок на 3D-печать МИЭМ НИУ ВШЭ.\\nАутентификация — OTP через email + JWT (Bearer token).",
        "title": "MIEM 3D Lab Print API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/api/admin/materials": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-materials"
                ],
                "summary": "Все материалы (admin)",
                "responses": {
                    "200": {
                        "description": "items",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_dto.MaterialResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-materials"
                ],
                "summary": "Создать материал (admin)",
                "parameters": [
                    {
                        "description": "Данные материала",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_dto.CreateMaterialRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_dto.MaterialResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Материал с таким названием уже существует",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/admin/materials/{id}": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-materials"
                ],
                "summary": "Обновить материал (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID материала",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Изменяемые поля",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_dto.PatchMaterialRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_dto.MaterialResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Название уже занято",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/admin/materials/{id}/colors": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-materials"
                ],
                "summary": "Добавить цвет к материалу (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID материала",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные цвета",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_dto.CreateColorRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_dto.ColorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Цвет с таким названием уже существует",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/admin/materials/{id}/colors/{color_id}": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-materials"
                ],
                "summary": "Обновить цвет (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID материала",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID цвета",
                        "name": "color_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Изменяемые поля",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_dto.PatchColorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_dto.ColorResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Название уже занято",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/materials": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "materials"
                ],
                "summary": "Список доступных материалов",
                "responses": {
                    "200": {
                        "description": "items",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_dto.MaterialResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Требуется согласие",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "miem_3d_lab_print_backend_src_internal_apierr.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "INTERNAL_ERROR"
                },
                "details": {
                    "type": "object"
                },
                "message": {
                    "type": "string",
                    "example": "Внутренняя ошибка сервера"
                }
            }
        },
        "miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.APIError"
                }
            }
        },
...
        "miem_3d_lab_print_backend_src_internal_dto.ColorResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "miem_3d_lab_print_backend_src_internal_dto.CreateColorRequest": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "miem_3d_lab_print_backend_src_internal_dto.CreateMaterialRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "miem_3d_lab_print_backend_src_internal_dto.MaterialResponse": {
            "type": "object",
            "properties": {
                "colors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_dto.ColorResponse"
                    }
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "miem_3d_lab_print_backend_src_internal_dto.PatchColorRequest": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "miem_3d_lab_print_backend_src_internal_dto.PatchMaterialRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите \"Bearer \u003caccess_token\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}